+++ src/.gitignore
@@ -1,5 +1,5 @@
-nmie/avx2/
-nmie/avx512/
+deepsparse/avx2/
+deepsparse/avx512/
 arch.bin
 cpu.py
 version.py
