+++ src/deepsparse/__init__.py
@@ -1,0 +1,8 @@
+"""
+The DeepSparse package used to achieve GPU class performance
+for Neural Networks on commodity CPUs.
+"""
+
+# flake8: noqa
+
+from .engine import *
