--- src/nmie/__init__.py
@@ -1,1 +1,0 @@
-from .model import *
